Add drf_cached_reads app
As part of bug 1063609, this app adds an instance caching strategy to
Django REST Framework, Instances are cached by model name and primary
key (PK), so that reading an instance can require 0 database reads, and
reading a list of instances takes 1 database read of the primary keys.

This app is intended to become a stand-alone project in the near future.
Test coverage is not complete, because it requires models and views to
exercise the code, which aren't appropriate until the project split.