bug 1063130 - Add canonical translation option
feature.name can either be a human-readable name or the "canonical" form
of the name, such as "<input>" for the HTML input element.  These are
stored in the database as the "zxx" translation (the ISO 639-2 "no
linguistic content" identifier), because JSONField doesn't store
strings.  They appear as strings in the JSON API output.