bug 1063141 - Import support data from webcompat
Import support data from the webcompat project.  By default, a subset of
the data is imported, due to lack of pagination on the list view.

There are several issues with the webcompat data, including not noting
the prefixes used by each version.  The importer will have to be
improved, or a new importer used, to correctly translate support
information from MDN.