bug 1063154 - Remove reverse relations
Removed most reverse relations from view_features in the linked
instances.  These are not needed for displaying an MDN table, and can
get quite large.  For example, you don't need to know the IDs of all the
versions or a browser, just the ones in the table, which are all
included in the linked.versions section.