fix bug 1063154 - Paginate deep feature trees
Large feature trees, such as the root feature node for CSS, can include
hundreds to thousands of descendant features, as well as an explosion of
related items.  This change detects a large feature tree and paginates
the view into 100-feature chunks.  For the CSS tree, it requires over
15 seconds and 1082 SQL queries to generate the first page on my laptop,
and 3.5 seconds and 2 queries from a warm cache.

This also caches the descendant IDs of small feature trees, for a slight
performance improvement in the more common case.