bug 1079920 - Keep section data when loading specs
When tools/load_spec_data.py was loading data from the kuma macros
SpecName and Spec2, it acted as if there were no related sections.  This
was OK, since you can't clear the section list by passing an empty list
of section IDs to the spec - this is used to re-order the sections in a
spec, and an empty list is ignored.  But, it did cause the diff to have
unexpected changes.

This change copies the related section IDs from the API server,
resulting in better diffs.

This also fixes downloads of unicode-encoded data, and moves cached data
to the data/ folder.