fix bug 1079920 - Unbreak feature tree
Adding subfeatures to the target feature also updates the tree structure
data through direct SQL. This is efficient, but the target feature's
tree structure data is reverted to the old values when it is eventually
saved. This change reads the tree data from the database and updates the
instance prior to saving the target feature.

Also, the backend is now ready for importing data from the English
versions of the pages, closing bug 1079920.  Additional tasks should be
captured in new tickets.