bug 1146430 - Remove exclude from serializer Meta
In DRF 2.4, fields and exclude can be used together to define what
fields appear in a serializer.  In DRF 3.1, either fields can be used
(define the included fields) or exclude (all model fields except
for...), but not both.  Added a helper method omit_some to help
construct the historical and view serializers, which omit some fields
from the standard serializers.