bug 1063565 - Add docs and integration testing
Adds scripts to generate representations of API requests and responses
for the purposes of documentation and integration testing.

tools/integration_requests.py runs in 3 modes:
- display - display HTTP requests and their results
- generate - Write HTTP headers an body for requests and results to
  docs/raw
- verify - Make HTTP requests, and compare to those in docs/raw

The requests are specified in docs/doc_cases.json.

docs/run_integration_tests.sh will:
- Create a fresh database, migrate it to the current version
- Create a test account
- Load resources from docs/resources
- Run the server
- Run tools/integration_requests.py against the server

When called normally, it acts as an integration test. When called with a
`-g` option, it re-generates the raw data to document API changes.