bug 1063565 - Standardize parameter setting
The command line arguments implied that the path to the raw folder
(containing standardized 'gold' runs for integration testing) could be
set with the --raw parameter.  However, this was not passed to the test
runner, but instead a relative path was always used.  This fixes this
bug, plus ensures that the CaseRunner and the command parser get their
default values from module variables.