bug 1171957 - Add mdn.utils.join_content
Same as mdn.scrape.PageVisitor.join_content, but without the
StripNextSpace hack needed because '3D' needs to be parsed one way for
compat features (text "3D") and another way for compat support
(version "3" followed by inline text "D").  This is the annoyance that
made this refactor seem like a good idea, so enjoy.