bug 1181140 - Add --reparse option to import_mdn
Previously, tools/import_mdn.py always did a reparse operation, which
reparses the cached MDN pages if available. This is the right option
when the scraper is updated and needs to be tested against the previous
scraper version. Now, --reparse is needed to get the old behaviour, and
the default is to redownload MDN pages and parse them. This is the right
option to periodically sync the API with MDN.