bug 1171957 - Whitelist HTML elements
Instead of blacklisting HTML elements that should not be allowed,
default to no elements allowed and whitelist the ones that are allowed.
This adds the tag_dropped issue for unexpected elements.

Also, handle a nested <table> inside a compatibility table, by moving
finalization to .to_feature_dict() and warning about <table>, <tr>, and
embedded <td> elements.