bug 1188498 - Switch to django-allauth branch
For a few views, django-allauth calls the template system in a way that
breaks when Django 1.8's Jinja2 backend is used.  This has been filed as
a bug [1], and the pull request accepted [2].  This code switches to the
current django-allauth master.

Another side effect of the Jinja2 backend is that assertTemplateUsed
always shows no templates were used. jingo got past this by rendering
templates manually and adding the appropriate signal handlers.  No one
has stepped up for Django yet [3].  This code replaces the test and
asserts the content has the correct title rather than asserts the
correct template was used.

[1] https://github.com/pennersr/django-allauth/pull/1133
[2] https://github.com/pennersr/django-allauth/pull/1134
[3] https://code.djangoproject.com/ticket/24622