fix bug 1128525 - Update remaining doc slugs
Update the following slugs in the docs:
* chrome_mobile -> android_webview
* chrome -> chrome_desktop

Chrome Mobile was an invalid class of browser - Chrome for Android was
probably what was meant.  Android Webview is a current browser that can
fill the same slot in the documentation.

The new slugs cause 4 browsers to be created in a different order, and
the related versions in a different order.  The doc resources and
results are updated to reflect the new browser and version IDs.

The view_feature examples now categorize all the documentation example
browsers as either desktop or mobile.