bug 1181158 - Stop caching non-English MDN pages
When the importer project started, one potential feature was scraping
localizations of feature names and notes from non-English pages. There
are several issues with this approach, including a lack of strong
mapping from English features to the same feature on the non-English
pages, and content issues fixed on the English page but present on the
non-English page. Instead, we'll focus our efforts on a contribution
interface for translators, and save the time and database space
downloading non-English MDN pages.