bug 1212865 - Set feature relation orders in tools
Treat feature.children as an ordered relationship. Create the parent
(w/o child IDs) before children, create children in sort order, and set
the order when writing to API.

Treat feature.sections as an order M2M relationship (the only one
currently in the API). Create the sections before the feature, create
the feature with the sections reference, and set the order when writing
to the API.

This changes the creation order of documentation resources, requiring
adjustments to IDs in documentation test cases:
- Version 10 (35.0) was accidentaly assigned to Browser 3 (Blackberry).
  Reassigned to Browser 4 (Chrome for Desktop).
- Feature 12 (web-css-transform-style) swapped with Feature 13
  (web-css-transform-three-value-syntax), changing many tests.
- The ID of some supports swapped, with no changes to tests

To help debug test changes, added a --stop option to
tools/integration_requests.py, and some commented-out code for
regenerating resources.