bug 1212865 - Fix update_cache_for_instance calls
Way back in b51bc9c, the update_only parameter was added to
update_cache_for_instance. With update_only as True, the cache
invalidation chain stops when the cache doesn't have an item, but when
False, a cached instance is created and the invalidation chain
continues. The intent was that update_only=False should be used on the
instance that was created or updated, and then controlled by
DRF_INSTANCE_CACHE_POPULATE_COLD. However, the code was accidently
setting version=False instead, leaving update_only True.

This commit fixes this, changing the default to update_only=False, rather
than using a positional argument.

Previously, newly created instances were not cached, but are now cached
as intended.  This means some unconverted datetime strings are passed to
cache serialization. drf_cached_instances 0.3.2 could not handle these,
but 0.3.3 can.

Some tests required updating, especially when new items were created in
a changeset.  This includes some API doc tests, which show that the
elements created in a changeset are now visible when the changeset is
closed.