bug 1224345 - Format subject feature errors
On PUT to view_feature, the subject feature could have a validation
error that isn't detected until the second validation phase, with the
linked resources. In this case, the resource sequence is None, and the
renderer needs to expect this.

This doesn't fix the second-phase validation issue, but it will allow
better debugging when it does occur again.