bug 1153288 - Refactor feature relation caching
Standardize the feature interface for Django instances and cached
instances, so that they are easier to use interchangably.

Previously, the cached feature instances would have additional
properties for relationships such as descendant features, implemented as
CachedQuerysets. ViewFeatureSerializer expected these relationships and
only worked with cached features.

This changes the properties, and implements them on the Django model
using the @cached_property decorator. This simplifies the cached feature
serializer, and means that either can be used in ViewFeatureSerializer.

The one difference is that cached features do not store very long
descendant lists. If this situation is detected, the Django feature
instance is loaded to be able to get the desendant PKs for pagination.