bug 1238587 - Ignore docstrings in tests
Docstrings have not been used for tests because they make nose output
less useful. When a test fails, it is useful to know the filename, test
class, and test method, so that the test can be run in isolation. So,
most tests have verbose names, and occasionally comments describing the
test purpose.

This adds a nose plugin to ignore test docstrings, so that they can be
used without the negative side effects. This may mean that PEP 257
warning D102 (missing docstring in public methods) could be re-enabled
in the future.