bug 1243217 - Require type member in parser
According to the JSON API v1.0 spec, "The type member is required in
every resource object throughout requests and responses" [1]. This
commit makes it a parse error to omit it. The desired error is a 409
Conflict [2], and the error format is wrong, but those are covered by
other bugs [3][4] and saved for the future.

[1] http://jsonapi.org/format/1.0/#crud
[2] http://jsonapi.org/format/1.0/#crud-creating-responses-409
[3] https://bugzilla.mozilla.org/show_bug.cgi?id=1243217
[4] https://bugzilla.mozilla.org/show_bug.cgi?id=1242649