bug 1238587 - Change bytes in initial migration
Local tests under Python 3.5 are having trouble with migration 0010
because the original default for this JSONField was the byte sequence
b'[]', and it needs to be the unicode string '[]'. That's the effect of
the migration, but it has to run under Python 2.7, which is more
flexible about what is valid JSON. So, hacking it by going back in time
and fixing the intial migration.

Resetting migrations will probably be part of the work on bug 1230306
(Convert to UUIDs) and bug 1153329 (renaming project to match
browsercompat), so this hack will soon be hidden in revision history.