bug 1230604 - Use tokens in integration tests
Generate and user OAuth2 tokens in integration tests. This impacts all
the request headers for tests requiring write access.

docs/v2/raw/changeset-related-user-response-body.json also changed, but
this is unexpected. The user instance cache is not being invalidated
when a changeset is created, and the cache is populated before the first
change when OAuth2 tokens are used. This is tracked in the new bug
1247690.