WIP bug 1230584 - Fix raw data form
The raw data form broke with recent Django REST Framework changes in how
non-POSTs are handled. This fixes it in DEBUG mode.

However, the code might not work in production, since the code is trying
to pull the CSRF token out of the headers. However, production uses
CSRF_COOKIE_HTTPONLY, which means JS can't access it. More testing, and
alternate code, is required.